fix(gidd)!: refuse invalid years, page sizes and release environments
A year was a Float, so a fractional endYear selected IDP stock with an
equality that matched nothing and published a zero. A negative pageSize
reached the ORM as a negative slice; its upper bound was a bare assert,
which an optimised build strips. An unrecognised releaseEnvironment fell
through to RELEASE and served release-year data with a 200.

The public statistics fields drop `required=True`: a non-null root field
propagates its null to `data`, so one bad argument emptied every sibling
field in the document.

+++ apps/gidd/tests/test_public_statistics.py
@@ -12,7 +12,7 @@ from utils.factories import (
 from utils.tests import HelixGraphQLTestCase
 
 CONFLICT_STATISTICS = """
-    query($clientId: String!, $startYear: Float, $endYear: Float) {
+    query($clientId: String!, $startYear: Int, $endYear: Int) {
         giddPublicConflictStatistics(clientId: $clientId, startYear: $startYear, endYear: $endYear) {
             newDisplacements
             totalDisplacements
@@ -25,7 +25,7 @@ CONFLICT_STATISTICS = """
 """
 
 DISASTER_STATISTICS = """
-    query($clientId: String!, $hazardTypes: [ID!], $startYear: Float, $endYear: Float) {
+    query($clientId: String!, $hazardTypes: [ID!], $startYear: Int, $endYear: Int) {
         giddPublicDisasterStatistics(
             clientId: $clientId, hazardTypes: $hazardTypes, startYear: $startYear, endYear: $endYear
         ) {
